+++ react/components/ReviewBlock.tsx
@@ -13,126 +13,17 @@ import {
   ToastContext,
 } from 'vtex.styleguide'
 import type { WrappedComponentProps } from 'react-intl'
-import { injectIntl, defineMessages } from 'react-intl'
+import { injectIntl } from 'react-intl'
 import PropTypes from 'prop-types'
 import { useApolloClient, useQuery } from 'react-apollo'
 
+import { reviewMessages as messages } from '../utils/messages'
 import { ParseText, GetText } from '../utils'
 import getRefIdTranslation from '../queries/refids.gql'
 import OrderFormQuery from '../queries/orderForm.gql'
+import autocomplete from '../queries/autocomplete.gql'
 
 const remove = <IconDelete />
-
-const messages = defineMessages({
-  valid: {
-    id: 'store/quickorder.valid',
-  },
-  available: {
-    id: 'store/quickorder.available',
-  },
-  invalidPattern: {
-    id: 'store/quickorder.invalidPattern',
-  },
-  withoutStock: {
-    id: 'store/quickorder.withoutStock',
-  },
-  skuNotFound: {
-    id: 'store/quickorder.skuNotFound',
-  },
-  withoutPriceFulfillment: {
-    id: 'store/quickorder.withoutPriceFulfillment',
-  },
-  cannotBeDelivered: {
-    id: 'store/quickorder.cannotBeDelivered',
-  },
-  ORD002: {
-    id: 'store/quickorder.ORD002',
-  },
-  ORD003: {
-    id: 'store/quickorder.ORD003',
-  },
-  ORD004: {
-    id: 'store/quickorder.ORD004',
-  },
-  ORD005: {
-    id: 'store/quickorder.ORD005',
-  },
-  ORD006: {
-    id: 'store/quickorder.ORD006',
-  },
-  ORD007: {
-    id: 'store/quickorder.ORD007',
-  },
-  ORD008: {
-    id: 'store/quickorder.ORD008',
-  },
-  ORD009: {
-    id: 'store/quickorder.ORD009',
-  },
-  ORD011: {
-    id: 'store/quickorder.ORD011',
-  },
-  ORD012: {
-    id: 'store/quickorder.ORD012',
-  },
-  ORD013: {
-    id: 'store/quickorder.ORD013',
-  },
-  ORD014: {
-    id: 'store/quickorder.ORD014',
-  },
-  ORD015: {
-    id: 'store/quickorder.ORD015',
-  },
-  ORD016: {
-    id: 'store/quickorder.ORD016',
-  },
-  ORD017: {
-    id: 'store/quickorder.ORD017',
-  },
-  ORD019: {
-    id: 'store/quickorder.ORD019',
-  },
-  ORD020: {
-    id: 'store/quickorder.ORD020',
-  },
-  ORD021: {
-    id: 'store/quickorder.ORD021',
-  },
-  ORD022: {
-    id: 'store/quickorder.ORD022',
-  },
-  ORD023: {
-    id: 'store/quickorder.ORD023',
-  },
-  ORD024: {
-    id: 'store/quickorder.ORD024',
-  },
-  ORD025: {
-    id: 'store/quickorder.ORD025',
-  },
-  ORD026: {
-    id: 'store/quickorder.ORD026',
-  },
-  ORD027: {
-    id: 'store/quickorder.ORD027',
-  },
-  ORD028: {
-    id: 'store/quickorder.ORD028',
-  },
-  ORD029: {
-    id: 'store/quickorder.ORD029',
-  },
-  ORD030: {
-    id: 'store/quickorder.ORD030',
-  },
-  ORD031: {
-    id: 'store/quickorder.ORD031',
-  },
-  cannotGetSkuInfo: {
-    id: 'store/quickorder.cannotGetSkuInfo',
-  },
-})
 
 let orderFormId = ''
 
@@ -153,6 +44,27 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     ssr: false,
     skip: !!orderFormId,
   })
+
+  const checkRestriction = async (sku: any) => {
+    return client.query({
+      query: autocomplete,
+      variables: { inputValue: sku },
+    })
+  }
+
+  const setRestriction = async (data: any) => {
+    return Promise.all(
+      data.map(async (item: any) => {
+        const res: any = await checkRestriction(item.refid)
+        if (
+          res?.data?.productSuggestions?.products[0]?.items[0]?.itemId ===
+          item.sku
+        ) {
+          return item
+        }
+      })
+    )
+  }
 
   const [state, setReviewState] = useState<any>({
     reviewItems:
@@ -178,6 +90,7 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     'store/quickorder.withoutStock': messages.withoutStock,
     'store/quickorder.withoutPriceFulfillment':
       messages.withoutPriceFulfillment,
+    'store/quickorder.limited': messages.limited,
     'store/quickorder.cannotBeDelivered': messages.cannotBeDelivered,
     'store/quickorder.ORD002': messages.ORD002,
     'store/quickorder.ORD003': messages.ORD003,
@@ -209,7 +122,7 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
     'store/quickorder.ORD031': messages.ORD031,
   }
 
-  const validateRefids = (refidData: any, reviewed: any) => {
+  const validateRefids = async (refidData: any, reviewed: any) => {
     let error = false
 
     // drops sellers without stock from refidData
@@ -245,7 +158,14 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
       const mappedRefId = {}
 
       if (refidData?.skuFromRefIds?.items) {
-        refidData.skuFromRefIds.items.forEach((item: any) => {
+        const restrictedData = await setRestriction(
+          refidData.skuFromRefIds.items
+        ).then((data) =>
+          data.filter((item: any) => {
+            return item != null
+          })
+        )
+        restrictedData.forEach((item: any) => {
           mappedRefId[item.refid] = item
         })
       }
@@ -305,7 +225,9 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
           totalQuantity: sellerUnitMultiplier
             ? sellerUnitMultiplier * item.quantity
             : '',
-          error: errorMsg(item),
+          error:
+            errorMsg(item) ??
+            (sellerWithStock ? null : `store/quickorder.limited`),
         }
       })
 
@@ -322,11 +244,13 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
       })
 
       onReviewItems(updated)
-      setReviewState({
-        ...state,
-        reviewItems: updated,
-        hasError: error,
-      })
+      if (JSON.stringify(reviewItems) !== JSON.stringify(updated)) {
+        setReviewState({
+          ...state,
+          reviewItems: updated,
+          hasError: error,
+        })
+      }
     }
   }